datadog/requirements.txt
datadog-api-client>=2.27,<3
python-dotenv>=1.0,<2
